refactor task_definition folder and add some comments
