+++ mobile-browser-based-version/src/helpers/README.md
@@ -10,6 +10,7 @@ This directory is the core of the backend of the platform. All helper javascript
 - **Memory**: handles loading and saving models and weights in memory
 - **Testing**: handles all tasks related to the testing of models (prediction, visualisation, etc.)
 - **Training**: handles all tasks related to the training of models (report loss values, create the live loss chart, etc.)
+- **Task definition**: handles all task related aspects. e.g. how to load them, preprocess them, make predictions and all of this for different data types (e.g. csv or image).
 
 ## Javascript ES6
